requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cassandra-driver==3.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-decouple==3.8
aiofiles==23.2.1
